Pull in changes from DxCore
The changes are no longer just relevant to non-megaAVR parts; features were added, bugs and a horriffic misfeature in the behavior of input as source (kept "input" behaving as it did, but added new "pin" to get what the behavior ought to be). Previously, it wasn't possible to have the pin be OUTPUT while using the value of that pin for the CCL.., but that's often your only (or perhaps, best) option to get LEVEL input from the code... software events don't latch, they're PULSE, and you can't change the LUT without stopping the whole CCL (thus introducing spikes on the outputs).... So having an OUTPUT pin be an input to CCL makes total sense.

Actually, it looks like you were already a version behind? You're missing one of the bugs too...