Correct description of feedback
feedback input takes the "output" of the *even* LUT (which is taken as the output of the sequencer if that is used, otherwise the output of the LUT itself. Unclear if intended and the documentation is poorly written. 

Also - were you going to cut the tinyAVR and Dx-stuff from the readme for MegaCoreX? (do note that I've cut the mega/Dx-stuff from megaTinyCore readme, and am cutting tiny stuff from DxCore version.