Move INADDR_NONE out of the arduino namespace to not break the Ethernet library
When I grafted the newer arduino namespace in for Modbus library
compatibility, I introduced a very subtle problem where constant objects
defined in the arduino namespace had to be fully qualified.

I suspect that the changes that arduino made to their API in the
versions since 1.0.0 somehow allowed for unqualified name resolution.